requirements.txt
asgiref==3.3.4
Django==3.2.4
django-ckeditor==6.1.0
django-js-asset==1.2.2
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
